pyproject.toml
[tool.poetry]
name = "passport-office"
version = "0.1.0"
description = ""
authors = ["Stas Soloviov <stas0601019@gmail.com>"]
readme = "README.md"
packages = [{include = "passport_office"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.2.3"
flask-sqlalchemy = "^3.0.3"
psycopg2-binary = "^2.9.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
